[CALCITE-3248] add Connector implementation. (Tino Rusch)
This is to support passing externally prepared HTTP clients for the jdbc connection.
It also includes the refactoring of the authentication methods to be separate
from the driver implementation, so that the user can setup the http client as
wanted and pass that in to the Connector. Also the unused username/password
fields in the dsn are removed.